Ban kick and mute

+++ Commands/prune.js
@@ -7,6 +7,8 @@ module.exports = class Prune {
   }
 
   async run(message, args) {
+    if(!message.member.hasPermission(`MANAGE_MESSAGES`)) return message.channel.sendMessage(`:warning: You required the **Manage Messages** permission.`);
+    if(!message.guild.member(this.client.user).hasPermission(`MANAGE_MESSAGES`)) return message.channel.sendMessage(`:warning: Hammer requires the **Manage Messages** permission.`);
     if(message.mentions.channels.first()) {
       message.mentions.channels.first().fetchMessages({ limit: parseInt(args.split(" ").pop()) }).then(messages => {
         message.mentions.channels.first().bulkDelete(messages).then(() => {
@@ -15,15 +17,17 @@ module.exports = class Prune {
       })
     } else if(message.mentions.users.first()) {
       message.channel.fetchMessages({ limit: 100 }).then(messages => {
-        let found = [];
-        messages.forEach(m => {
-          if(m == message.mentions.users.first().id)
-          found.push(m);
-          message.channel.sendMessage(m);
+        let count = parseInt(args.split(" ").pop()) || 5
+        let found = []
+        messages.array().forEach(m => {
+          if(m.author.id == message.mentions.users.first().id) {
+            if(found.length == count) return;
+            found.push({ id: m.id })
+          }
         })
         message.channel.bulkDelete(found).then(() => {
-          message.channel.sendMessage(`:ok_hand: Pruned **${found.length}** messages from this channel.`);
-        }).catch(e => message.channel.sendMessage(`:x: Failed to prune **${found.length}** messages from this channel.\n${e}`))
+          message.channel.sendMessage(`:ok_hand: Pruned **${found.length}** messages from <@${message.mentions.users.first().id}>.`);
+        }).catch(e => { message.channel.sendMessage(`:x: Failed to prune **${found.length}** messages from <@${message.mentions.users.first().id}>.\n${e}`); console.error(e); })
       })
     } else {
       message.channel.fetchMessages({ limit: parseInt(args) }).then(messages => {
